JobDashboard.py
import csv
from jobspy import scrape_jobs

jobs = scrape_jobs(
    site_name=["indeed", "linkedin", "glassdoor"], # "glassdoor", "bayt", "naukri", "bdjobs"
    search_term='UX',
    location="Edinburgh",
    results_wanted=30,
    hours_old=168,
    country_indeed="UK",
    # is_remote=False,
    
    # linkedin_fetch_description=True # gets more info such as description, direct job url (slower)
    # proxies=["208.195.175.46:65095", "208.195.175.45:65095", "localhost"],
)
print(f"Found {len(jobs)} jobs")
print(jobs.head())
jobs.to_csv("jobs.csv", quoting=csv.QUOTE_NONNUMERIC, escapechar="\\", index=False) # to_excel