pyproject.toml
[project]
name = "jobspy"
version = "0.1.0"
description = ""
authors = [
    {name = "Job Hunter",email = "user@example.com"}
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "python-jobspy>=1.1.82"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
